+++ rca_app/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.8.10"
-RCA_CORE_VERSION = "0.8.9"
+__version__ = "1.8.11"
+RCA_CORE_VERSION = "0.8.10"